Rename document color collection to follow naming path from other places.
